canvas max width support
