+++ packages/app/src/library/Nominations/index.tsx
@@ -123,7 +123,6 @@ export const Nominations = ({
                     nominator,
                     nominated,
                   },
-                  size: 'xl',
                 })
               }
             />
