+++ packages/app/src/pages/Nominate/Active/CommissionPrompt.tsx
@@ -62,7 +62,6 @@ export const CommissionPrompt = () => {
                     nominator: activeAccount,
                     nominated,
                   },
-                  size: 'xl',
                 })
               }
             />
