+++ packages/ui-core/src/canvas/Main/index.module.scss
@@ -1,11 +1,21 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+@import '../_variables';
+
 .main {
   padding-top: 3rem;
   min-height: calc(100vh - 12rem);
   padding-bottom: 2rem;
   width: 100%;
+
+  &.lg {
+    max-width: $canvas-width-lg;
+  }
+  &.xl {
+    max-width: $canvas-width-xl;
+  }
+  
   > h1 {
     margin-top: 1.5rem;
     margin-bottom: 1.25rem;
