+++ packages/ui-core/src/canvas/_variables.scss
@@ -1,0 +1,5 @@
+/* Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+SPDX-License-Identifier: GPL-3.0-only */
+
+$canvas-width-lg: 800px;
+$canvas-width-xl: 1250px;
